refactor: consolidate UI components by moving CliModel and CliProvider to model module
